+++ packages/mdx/bin.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-import { postInstall } from './dist/next/index.js';
-
-void postInstall(process.argv[2], process.argv[3]);