+++ packages/mdx/src/mdx/build-mdx.ts
@@ -1,11 +1,15 @@
 import { createProcessor, type ProcessorOptions } from '@mdx-js/mdx';
-import type { DataMap, VFile } from 'vfile';
-import { remarkInclude } from '@/mdx-plugins/remark-include';
+import type { VFile } from 'vfile';
+import { remarkInclude } from '@/mdx/remark-include';
 import type { StructuredData } from 'fumadocs-core/mdx-plugins';
 import type { TableOfContents } from 'fumadocs-core/server';
 import type { FC } from 'react';
 import type { MDXProps } from 'mdx/types';
-import type { ExtractedReference } from '@/mdx-plugins/remark-postprocess';
+import {
+  type ExtractedReference,
+  type PostprocessOptions,
+  remarkPostprocess,
+} from '@/mdx/remark-postprocess';
 
 type Processor = ReturnType<typeof createProcessor>;
 
@@ -25,6 +29,7 @@ interface BuildMDXOptions extends ProcessorOptions {
   data?: Record<string, unknown>;
 
   _compiler?: CompilerOptions;
+  postprocess?: PostprocessOptions;
 }
 
 export interface CompilerOptions {
@@ -42,24 +47,38 @@ export interface CompiledMDXProperties<Frontmatter = Record<string, unknown>> {
    */
   lastModified?: Date;
   extractedReferences?: ExtractedReference[];
+  _markdown?: string;
 }
 
-export type { DataMap };
+export interface FumadocsDataMap {
+  /**
+   * [Fumadocs MDX] raw frontmatter, you can modify it
+   */
+  frontmatter?: Record<string, unknown>;
+
+  /**
+   * [Fumadocs MDX] additional ESM exports to write
+   */
+  'mdx-export'?: { name: string; value: unknown }[];
+
+  extractedReferences: ExtractedReference[];
+
+  /**
+   * [Fumadocs MDX] The compiler object from loader
+   */
+  _compiler?: CompilerOptions;
+
+  _getProcessor?: (format: 'md' | 'mdx') => Processor;
+
+  /**
+   * [Fumadocs MDX] Processed Markdown content before `remark-rehype`.
+   */
+  _markdown?: string;
+}
 
 declare module 'vfile' {
-  interface DataMap {
-    /**
-     * [Fumadocs MDX] raw frontmatter, you can modify it
-     */
-    frontmatter?: Record<string, unknown>;
-
-    /**
-     * [Fumadocs MDX] The compiler object from loader
-     */
-    _compiler?: CompilerOptions;
-
-    _getProcessor?: (format: 'md' | 'mdx') => Processor;
-  }
+  // eslint-disable-next-line @typescript-eslint/no-empty-object-type -- extend data map
+  interface DataMap extends FumadocsDataMap {}
 }
 
 /**
@@ -82,7 +101,24 @@ export async function buildMDX(
       processor = createProcessor({
         outputFormat: 'program',
         ...rest,
-        remarkPlugins: [remarkInclude, ...(rest.remarkPlugins ?? [])],
+        remarkPlugins: [
+          remarkInclude,
+          ...(rest.remarkPlugins ?? []),
+          [
+            remarkPostprocess,
+            {
+              ...options.postprocess,
+              valueToExport: [
+                ...(options.postprocess?.valueToExport ?? []),
+                'structuredData',
+                'extractedReferences',
+                'frontmatter',
+                'lastModified',
+                '_markdown',
+              ],
+            } satisfies PostprocessOptions,
+          ],
+        ],
         format,
       });