+++ packages/mdx/src/runtime/next/types.ts
@@ -2,24 +2,12 @@ import type { StandardSchemaV1 } from '@standard-schema/spec';
 import type { MetaData, PageData, Source } from 'fumadocs-core/source';
 import type { LoadedConfig } from '@/utils/config';
 import type { DocCollection, DocsCollection, MetaCollection } from '@/config';
-import type { CompiledMDXProperties } from '@/utils/build-mdx';
-
-export interface BaseCollectionEntry {
-  /**
-   * Raw file path of collection entry, including absolute path (not normalized).
-   */
-  _file: FileInfo;
-}
-
-export interface FileInfo {
-  path: string;
-  absolutePath: string;
-}
-
-interface MarkdownProps
-  extends Omit<CompiledMDXProperties, 'frontmatter' | 'default'> {
-  body: CompiledMDXProperties['default'];
-}
+import type {
+  AsyncDocCollectionEntry,
+  DocCollectionEntry,
+  FileInfo,
+  MetaCollectionEntry,
+} from '@/runtime/shared';
 
 export interface RuntimeFile {
   info: FileInfo;
@@ -33,25 +21,11 @@ export interface AsyncRuntimeFile {
   lastModified?: Date;
 }
 
-type DocOut<Schema extends StandardSchemaV1> = Override<
-  MarkdownProps & {
-    /**
-     * Other exports in the compiled Markdown/MDX file
-     */
-    _exports: Record<string, unknown>;
+export type DocOut<Schema extends StandardSchemaV1 = StandardSchemaV1> =
+  DocCollectionEntry<StandardSchemaV1.InferOutput<Schema>>;
 
-    /**
-     * Read the original content of file from file system.
-     */
-    get content(): string;
-  },
-  StandardSchemaV1.InferOutput<Schema> & BaseCollectionEntry
->;
-
-type Override<A, B> = Omit<A, keyof B> & B;
-
-type MetaOut<Schema extends StandardSchemaV1> =
-  StandardSchemaV1.InferOutput<Schema> & BaseCollectionEntry;
+export type MetaOut<Schema extends StandardSchemaV1 = StandardSchemaV1> =
+  MetaCollectionEntry<StandardSchemaV1.InferOutput<Schema>>;
 
 export interface Runtime {
   doc: <C>(
@@ -80,12 +54,8 @@ export interface Runtime {
     : never;
 }
 
-type AsyncDocOut<Schema extends StandardSchemaV1> =
-  StandardSchemaV1.InferOutput<Schema> &
-    BaseCollectionEntry & {
-      content: string;
-      load: () => Promise<MarkdownProps & Record<string, unknown>>;
-    };
+export type AsyncDocOut<Schema extends StandardSchemaV1 = StandardSchemaV1> =
+  AsyncDocCollectionEntry<StandardSchemaV1.InferOutput<Schema>>;
 
 export interface RuntimeAsync {
   doc: <C>(