+++ packages/mdx/tsup.config.ts
@@ -2,11 +2,10 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig({
   entry: [
-    './src/{index,loader-mdx}.ts',
+    './src/{index,loader-mdx,bin}.ts',
     './src/{config,next,vite,bun}/index.ts',
     './src/node/loader.ts',
-    './src/config/zod-3.ts',
-    './src/runtime/async.ts',
+    './src/runtime/next/{index,async}.ts',
     './src/runtime/vite/{browser,server}.ts',
   ],
   format: ['esm', 'cjs'],